+++ ase/io/ulm.py
@@ -126,6 +126,7 @@ from typing import Union, Set
 
 import numpy as np
 
+from ase.io.formats import is_compressed
 from ase.io.jsonio import encode, decode
 from ase.utils import plural
 
@@ -648,7 +649,7 @@ class NDArrayReader:
         offset = self.offset + start * self.itemsize * stride
         self.fd.seek(offset)
         count = (stop - start) * stride
-        if self.hasfileno:
+        if not is_compressed(self.fd) and self.hasfileno:
             a = np.fromfile(self.fd, self.dtype, count)
         else:
             # Not as fast, but works for reading from tar-files: